feat: BottomSheet Enum 분리 (#25)
