+++ BBANGZIP/BBANGZIP/Sources/Presentation/Component/BottomSheetType.swift
@@ -1,0 +1,21 @@
+//
+//  BottomSheetType.swift
+//  BBANGZIP
+//
+//  Created by 송여경 on 1/16/25.
+//  Copyright © 2025 com.bbangzip. All rights reserved.
+//
+
+import SwiftUI
+
+enum BottomSheetType: Int {
+    case revert
+    case sort
+    case withdraw
+    case detailBadge
+    case congratsBadge
+    case examDate
+    case studyFinishDate
+    case changeSemester
+}
+
